backend/lambdas/usage-tracker/requirements.txt
psycopg2-binary>=2.9.9
boto3>=1.34.0
stripe>=9.0.0
